requirements.txt
redis>=5.0.0
python-dotenv>=1.0.0
click>=8.1.0
sentence-transformers>=2.2.0
numpy>=1.21.0